+++ connectionstatistics/src/main/java/com/techolution/mauritius/smartwater/connection/domain/Telemetry.java
@@ -6,10 +6,10 @@ import java.util.Date;
 public class Telemetry implements Serializable {
 	
 	
-	public Long getReading() {
+	public Double getReading() {
 		return reading;
 	}
-	public void setReading(Long reading) {
+	public void setReading(Double reading) {
 		this.reading = reading;
 	}
 	/**
@@ -17,10 +17,10 @@ public class Telemetry implements Serializable {
 	 */
 	private static final long serialVersionUID = 1L;
 	private int meter_id;
-	private Long flow;
+	private Double flow;
 	private Long battery;
 	private Long flowrate;
-	private Long reading;
+	private Double reading;
 	private Date date;
 	public int getMeter_id() {
 		return meter_id;
@@ -29,10 +29,10 @@ public class Telemetry implements Serializable {
 		this.meter_id = meter_id;
 	}
 	
-	public Long getFlow() {
+	public Double getFlow() {
 		return flow;
 	}
-	public void setFlow(Long flow) {
+	public void setFlow(Double flow) {
 		this.flow = flow;
 	}
 	public Long getBattery() {
